Add authentication to MongoHelper and fix its package name
