+++ RespondaSePuder/grails-app/controllers/br/ufscar/sead/loa/respondasepuder/remar/QuestionController.groovy
@@ -1,7 +1,7 @@
 package br.ufscar.sead.loa.respondasepuder.remar
 
 import br.ufscar.sead.loa.remar.User
-import br.ufscar.sead.load.remar.api.MongoHelper
+import br.ufscar.sead.loa.remar.api.MongoHelper
 import grails.util.Environment
 import grails.plugin.springsecurity.annotation.Secured
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
